5-agent-rag/requirements.txt
langchain==0.3.19
langchain-openai==0.3.7
langchain-community==0.3.18
langchain-core==0.3.40
langchain-text-splitters==0.3.6
python-dotenv>=1.0.0
streamlit>=1.32.0
openai>=1.12.0
typing-extensions>=4.5.0
langchain_chroma>=0.1.2
colorama>=0.4.6
requests>=2.31.0
supabase>=1.1.0
pypdf>=3.15.0
langchain_community>=0.3.18
